requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
dask==2021.5.0
fastapi==0.65.0
Flask==1.1.2
fsspec==2021.5.0
h11==0.12.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
locket==0.2.1
MarkupSafe==1.1.1
mo-future==3.51.20059
moz-sql-parser==3.32.20026
numpy==1.20.2
pandas==1.2.4
partd==1.2.0
pyarrow==4.0.0
pycodestyle==2.7.0
pydantic==1.8.1
pyparsing==2.3.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.19.1
six==1.15.0
starlette==0.14.2
toolz==0.11.1
typing-extensions==3.10.0.0
uvicorn==0.13.4
urllib3==1.26.5
Werkzeug==1.0.1
